wt88047: camera: AOSP compatible HAL and update
* Adapted from rendang Marshmallow review
* Update qomx_core and wrapper
* Don't too much, our KitKat blobs not always happy

+++ camera/mm-image-codec/qomx_core/QOMX_JpegExtensions.h
@@ -273,7 +273,7 @@ typedef struct {
   void *handle;
   void *mem_hdl;
   int8_t isheap;
-  int size; /*input*/
+  size_t size; /*input*/
   void *vaddr;
   int fd;
 } omx_jpeg_ouput_buf_t;
